Intrested in starting a program to help others
Looking into starting a sort if program or organization if able to go through the steps to creating a way to reach out to the people struggling with the means to be able to stand up for themselves and loved ones best interest by a variety of different circumstances that with enough determination and support can get the opportunity to get the past circumstances holding them back no longer stop them from being given the chance to be the right thing and after the  capability to pass it forward and if we all keep finding new ways to help others, just maybe this world would become just a little less disappointing in general and some what down right evil. Maybe JUST maybe more people would see how amazing it feels to help others and to do to them what might not have been given so freely for every human being that doesn't go around hurting others and does the next right thing I want to make sure everyone possible can at least get the respect of a fair chance to be human and learn from the mistakes everyone makes. Making it possible to possibly change a person or families entire live with just a matter of not even that much effort. The possibilities are endless if people united and work together to lift up other and lending a hand to try and people and little by little remove more and more of the bad and replacing it with more good in this world.